+++ src/FastRoute.php
@@ -5,11 +5,11 @@ namespace Middlewares;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\ResponseInterface;
 use Interop\Container\ContainerInterface;
-use Interop\Http\Middleware\ServerMiddlewareInterface;
-use Interop\Http\Middleware\DelegateInterface;
+use Interop\Http\ServerMiddleware\MiddlewareInterface;
+use Interop\Http\ServerMiddleware\DelegateInterface;
 use FastRoute\Dispatcher;
 
-class FastRoute implements ServerMiddlewareInterface
+class FastRoute implements MiddlewareInterface
 {
     /**
      * @var Dispatcher FastRoute dispatcher